Prepped Autoloader for Namespaces
Removed the "loadAll" function, changed the class names to the hand
less to support the new namespaced format, and added a "psr_autoload"
function to help with the transition (so I can test along the way).
Ultimately the psr_autoload (or something similar) will end up being
the main autoload function.