Changed the key structure and added a getKey function
Updated the setupKey function to no longer take unlimited arguments,
but to instead take either a single string representing a path
(path/to/cache) or an array representing the same. This doesn't change
the functionality much, but does simplify things in many ways. Because
this eliminates some arguments it means some error conditions were no
longer needed.

Also added the "getKey" function, which will return the key as a string
(even if passed as an array).