Removed ability to create Item without a key.
Creating an Item without a key was an artifact from the pre-Pool days,
where certain actions (such as clearing) needed to be done rather than
actions on single items. Now that the pool exists to handle these it
makes little sense for the Item to be able to exist without a key.