Fix memory leak caused by static query logging
This commit resolves #79 by removing all query logging from the Item class.
In regars to seperation of concern and as pointed out by @jhallbachner this
is not the place to put such functionality, and neither does it have to be
at this level as it can be done in a layer outside.