Split constructor arguments out, removed "string" key
The constructor is no longer there- instead there is a setDriver and
setKey function.

The Item class no longer accepts a string for a key, it must be
provided with an array. This was done for efficiency (since the Pool
class already converts it to an array) and to make namespacing easier.

This does break previous API compatibility, however it should have
minimal affects as these functions are all used internally (typically
by the Pool class when generating the Item).