adds tests for RedisArray options
In order to test whether RedisArray options are correctly passed, use
uopz extension which allows to stub the constructor of PHP objects so we
can test the parameter passing without dependency-injection and factory
hell. It also keeps the API compatible because otherwise we would have
to pass e. g. factory.