Created stand alone "save" function that uses the data and ttl properties
This is because in PSR6 the “set” and new expiration functions are no
longer persistent to the driver but instead save data in the object
until it is saved.

This change provides backwards compatibility (for now) to make testing
easier, while providing the foundation of the future changes to this
class.