Add support for connection persistence to the memcached driver
When using PHP Memcached there's an optional parameter called
$persistent_id (see http://php.net/manual/en/memcached.construct.php.

This change adds support for that option by adding it as a parameter
to the array of options passed into Memcache::setOptions which is
then passed down to the SubMemcached.

If not set then null is used and persistence is disabled.